backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["PandorasActor0 <nelsona2001@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
numpy = "^1.26.0"
pandas = "^2.1.1"
matplotlib = "^3.8.0"
tensorflow-hub = "^0.14.0"
seaborn = "^0.13.0"
nltk = "^3.8.1"
scikit-learn = "^1.3.1"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
